orchestration/requirements.txt
apache-airflow==3.0.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pyspark==3.5.6
apache-airflow-providers-apache-spark==5.3.0
apache-airflow-providers-amazon==9.8.0
apache-airflow-providers-postgres==6.2.0
kafka-python==2.2.10
